backend/requirements.txt
# requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
websockets==11.0.3
requests==2.31.0
openai-whisper==20230314
TTS==0.21.3